Refactor status API and pages structure
