+++ crates/hotfix-status/src/ui.rs
@@ -1,20 +1,13 @@
-use askama::Template;
+use crate::ui::assets::static_handler;
+use crate::ui::dashboard::dashboard_handler;
 use axum::Router;
-use axum::response::{Html, IntoResponse};
 use axum::routing::get;
 
-#[derive(Template)]
-#[template(path = "dashboard.askama")]
-struct DashboardTemplate<'a> {
-    title: &'a str,
-}
-
-async fn dashboard_handler() -> impl IntoResponse {
-    let template = DashboardTemplate { title: "Dashboard" };
-
-    Html(template.render().unwrap())
-}
+mod assets;
+mod dashboard;
 
 pub fn builder_ui_router() -> Router {
-    Router::new().route("/dashboard", get(dashboard_handler))
+    Router::new()
+        .route("/dashboard", get(dashboard_handler))
+        .route("/static/{*file}", get(static_handler))
 }
