+++ crates/hotfix-status/src/ui/assets.rs
