+++ crates/hotfix-status/src/ui/dashboard.rs
@@ -1,0 +1,14 @@
+use askama::Template;
+use axum::response::{Html, IntoResponse};
+
+#[derive(Template)]
+#[template(path = "dashboard.askama")]
+struct DashboardTemplate<'a> {
+    title: &'a str,
+}
+
+pub(crate) async fn dashboard_handler() -> impl IntoResponse {
+    let template = DashboardTemplate { title: "Dashboard" };
+
+    Html(template.render().unwrap())
+}
