+++ crates/hotfix-status/templates/dashboard.askama
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <title>{{ title }}</title>
-    <script src="/assets/tailwind.js"></script>
+    <script src="/static/tailwind.js"></script>
 </head>
 <body>
 <div class="bg-gray-500">
