+++ examples/simple-new-order/src/main.rs
@@ -147,6 +147,6 @@ async fn start_status_service(session_ref: SessionRef<Message>) {
     let host_and_port = std::env::var("HOST_AND_PORT").unwrap_or("0.0.0.0:9881".to_string());
     let listener = tokio::net::TcpListener::bind(&host_and_port).await.unwrap();
 
-    info!("starting status service on {host_and_port}");
+    info!("starting status service on http://{host_and_port}");
     axum::serve(listener, status_router).await.unwrap();
 }
